requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pyyaml==6.0.2

SQLAlchemy==2.0.32
asyncpg==0.29.0
minio==7.2.7
yt-dlp==2024.08.06
python-multipart==0.0.9
youtube-transcript-api==0.6.2
openai==1.40.2
